# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: LidRadar API
  version: 1.0.0
  description: >-
    Целевой контракт MVP. Пути без расширения x-lidradar-runtime-status уже
    подключены к рабочему API. Значение planned означает, что контракт
    зафиксирован для будущего этапа, но маршрут ещё не доступен в cmd/api.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /
    description: Current LidRadar API origin
security:
  - sessionCookie: []
paths:
  /internal/v1/ai/nodes/heartbeat:
    post:
      summary: Renew AI node readiness and leases owned by this node
      operationId: heartbeatAINode
      security: [{ aiNodeBearer: [] }]
      parameters:
        - { $ref: '#/components/parameters/AINodeId' }
        - { $ref: '#/components/parameters/AITimestamp' }
        - { $ref: '#/components/parameters/AIRequestId' }
        - { $ref: '#/components/parameters/AISignature' }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/AINodeHeartbeatRequest' } } }
      responses:
        '204': { description: Heartbeat accepted and current leases renewed. }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/AINodeUnauthorized' }
  /internal/v1/ai/jobs/claim:
    post:
      summary: Atomically lease one available AI job
      description: Returns prompts only for tenants explicitly assigned to the authenticated node and only when the model version matches.
      operationId: claimAIJob
      security: [{ aiNodeBearer: [] }]
      parameters:
        - { $ref: '#/components/parameters/AINodeId' }
        - { $ref: '#/components/parameters/AITimestamp' }
        - { $ref: '#/components/parameters/AIRequestId' }
        - { $ref: '#/components/parameters/AISignature' }
      responses:
        '200': { description: One job leased to the authenticated node., content: { application/json: { schema: { $ref: '#/components/schemas/AIJobClaim' } } } }
        '204': { description: No compatible job is currently available. }
        '401': { $ref: '#/components/responses/AINodeUnauthorized' }
...
  /api/v1/risks/{riskId}/recommendation:
    parameters: [{ $ref: '#/components/parameters/TenantId' }, { $ref: '#/components/parameters/RiskId' }]
    post:
      summary: Создать или получить детерминированную рекомендацию по риску
      description: Тип риска читается сервером из PostgreSQL; команда не обращается к AI.
      operationId: ensureRiskRecommendation
      responses:
        '200': { description: Созданная ранее либо новая шаблонная рекомендация., content: { application/json: { schema: { $ref: '#/components/schemas/Recommendation' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/risks/{riskId}/actions:
    parameters: [{ $ref: '#/components/parameters/TenantId' }, { $ref: '#/components/parameters/RiskId' }, { $ref: '#/components/parameters/IdempotencyKey' }]
    post:
      summary: Записать неизменяемое корректирующее действие
      operationId: createRiskAction
      requestBody: { required: true, content: { application/json: { schema: { type: object, additionalProperties: false, required: [type], properties: { type: { $ref: '#/components/schemas/ActionType' }, note: { type: string, maxLength: 2000 } } } } } }
      responses:
        '200': { description: Ранее сохранённый ответ при точном повторе., content: { application/json: { schema: { $ref: '#/components/schemas/Action' } } } }
        '201': { description: Неизменяемое действие создано вместе с аудитом., content: { application/json: { schema: { $ref: '#/components/schemas/Action' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { description: Ключ идемпотентности использован с другим содержимым., content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /api/v1/opportunities/{opportunityId}/outcomes:
    parameters:
      - { $ref: '#/components/parameters/TenantId' }
      - { $ref: '#/components/parameters/OpportunityId' }
      - { $ref: '#/components/parameters/IdempotencyKey' }
    post:
      summary: Записать неизменяемый исход коммерческой возможности
      operationId: createOpportunityOutcome
      requestBody: { required: true, content: { application/json: { schema: { type: object, additionalProperties: false, required: [status], properties: { status: { $ref: '#/components/schemas/OutcomeStatus' }, note: { type: string, maxLength: 2000 } } } } } }
      responses:
        '200': { description: Ранее сохранённый ответ при точном повторе., content: { application/json: { schema: { $ref: '#/components/schemas/Outcome' } } } }
        '201': { description: Неизменяемый исход создан вместе с аудитом., content: { application/json: { schema: { $ref: '#/components/schemas/Outcome' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { description: Ключ идемпотентности использован с другим содержимым., content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: lidradar_session
    aiNodeBearer:
      type: http
      scheme: bearer
      bearerFormat: LidRadar AI node secret
  parameters:
    TenantId: { name: X-Tenant-ID, in: header, required: true, description: Explicit Organization tenant selected from /auth/me memberships., schema: { type: string, format: uuid } }
...
    OpportunityId: { name: opportunityId, in: path, required: true, schema: { type: string, format: uuid } }
...
    RiskId: { name: riskId, in: path, required: true, schema: { type: string, format: uuid } }
...
    IdempotencyKey: { name: Idempotency-Key, in: header, required: true, schema: { type: string, minLength: 1, maxLength: 255 } }
    AINodeId: { name: X-LidRadar-Node-ID, in: header, required: true, schema: { type: string, format: uuid } }
    AITimestamp: { name: X-LidRadar-Timestamp, in: header, required: true, description: UTC RFC3339Nano timestamp inside the configured replay window., schema: { type: string, format: date-time } }
    AIRequestId: { name: X-LidRadar-Request-ID, in: header, required: true, description: Unique request nonce retained only for the replay window., schema: { type: string, format: uuid } }
    AISignature: { name: X-LidRadar-Signature, in: header, required: true, description: 'HMAC-SHA256 over method, path, timestamp, request ID, and SHA-256 body hash.', schema: { type: string, pattern: '^sha256=[0-9a-f]{64}$' } }
...
  responses:
    BadRequest: { description: 'Request body, tenant selector, or domain value is invalid.', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Unauthorized: { description: 'Opaque session cookie is missing, expired, revoked, or invalid.', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
    Forbidden: { description: Request origin is not trusted or Membership does not grant the required named permission., content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    NotFound: { description: Resource does not exist in the authenticated tenant, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
    AINodeUnauthorized: { description: 'AI node credentials, timestamp, nonce, body hash, or request signature are invalid.', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
  schemas:
    AINodeHeartbeatRequest:
      type: object
      additionalProperties: false
      required: [status, modelVersion, availableSlots]
      properties:
        status: { type: string, enum: [READY, OFFLINE] }
        modelVersion: { type: string, minLength: 1, maxLength: 200 }
        availableSlots: { type: integer, minimum: 0, maximum: 1 }
    AIJobClaim:
      type: object
      additionalProperties: false
      required: [id, tenantId, jobType, entityType, entityId, prompt, baseConversationRevision, analysisThroughMessageId, modelVersion, promptVersion, schemaVersion, leaseUntil]
      properties:
        id: { type: string, format: uuid }
        tenantId: { type: string, format: uuid }
        jobType: { type: string, enum: [ANALYZE_CONVERSATION] }
        entityType: { type: string, enum: [CONVERSATION] }
        entityId: { type: string, format: uuid }
        prompt: { type: string, minLength: 1, writeOnly: true }
        baseConversationRevision: { type: integer, format: int64, minimum: 1 }
        analysisThroughMessageId: { type: string, format: uuid }
        modelVersion: { type: string }
        promptVersion: { type: string }
        schemaVersion: { type: string }
        leaseUntil: { type: string, format: date-time }
...
    Error:
      type: object
      required: [error]
      properties: { error: { type: object, required: [code, message, traceId], properties: { code: { type: string }, message: { type: string }, details: { type: object, additionalProperties: true }, traceId: { type: string } } } }
...
    ActionType: { type: string, enum: [OPEN_CONVERSATION, COPY_REPLY, MARK_CONTACTED, CALL, SEND_MESSAGE, OTHER] }
    OutcomeStatus: { type: string, enum: [RESPONDED, BOOKED, PAID, LOST, THINKING, NOT_A_LEAD] }
    Recommendation:
      type: object
      additionalProperties: false
      required: [id, riskId, text, source, createdAt]
      properties: { id: { type: string, format: uuid }, riskId: { type: string, format: uuid }, text: { type: string, minLength: 1, maxLength: 2000 }, source: { type: string, enum: [TEMPLATE] }, createdAt: { type: string, format: date-time } }
    Action:
      type: object
      additionalProperties: false
      required: [id, riskId, actorId, type, createdAt]
      properties: { id: { type: string, format: uuid }, riskId: { type: string, format: uuid }, actorId: { type: string, format: uuid }, type: { $ref: '#/components/schemas/ActionType' }, note: { type: string, maxLength: 2000 }, createdAt: { type: string, format: date-time } }
    Outcome:
      type: object
      additionalProperties: false
      required: [id, opportunityId, actorId, status, createdAt]
      properties: { id: { type: string, format: uuid }, opportunityId: { type: string, format: uuid }, actorId: { type: string, format: uuid }, status: { $ref: '#/components/schemas/OutcomeStatus' }, note: { type: string, maxLength: 2000 }, createdAt: { type: string, format: date-time } }
...